refactor: все sql запросы возвращены в хранилища (репозитории), удалены повторные проверки на ошибки в различных методах UserService и FilmService. delete: удалены реализации в памяти UserStorage и FilmStorage.
