+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -13,4 +13,12 @@ public interface UserStorage {
     User update(User user);
 
     User getUser(int id);
+
+    void addFriend(int id, int friendId);
+
+    void deleteFriend(int id, int friendId);
+
+    List<User> getFriends(int id);
+
+    List<User> getMutualFriends(int id, int friendId);
 }
